Fixed errors of hinter and adding files
